quant analyse tools file structure change; quant analyse tools adapt fp8 scale type change
